+++ apps/web/src/entities/expense/index.ts
@@ -1,4 +1,1 @@
-export { AmountInput } from './ui';
-export type { AmountInputProps } from './ui';
-
 export { EXPENSE_CONSTANTS, EXPENSE_ERROR_MESSAGES } from './config/constants';
