--- apps/web/src/entities/expense/ui/AmountInput/index.ts
@@ -1,2 +1,0 @@
-export { AmountInput } from './AmountInput';
-export type { AmountInputProps } from './AmountInput';
