+++ apps/web/src/features/expense/ui/AmountInput/index.ts
