+++ apps/web/src/features/expense/ui/index.ts
@@ -1,3 +1,6 @@
+export { AmountInput } from './AmountInput';
+export type { AmountInputProps } from './AmountInput';
+
 export {
   ExpenseFormBottomSheet,
   CategoryBottomSheetTemplate,
