+++ apps/web/src/features/expense/ui/steps/AmountDateStep/AmountDateStep.tsx
@@ -6,7 +6,7 @@ import { CalendarBottomSheetTemplate } from './CalendarBottomSheet';
 import { useModal } from '@/shared/hooks';
 import * as styles from './AmountDateStep.css';
 import { formatDate, formatNumberWithComma } from '@/shared/utils';
-import { AmountInput } from '@/entities/expense';
+import { AmountInput } from '../../AmountInput';
 
 export interface AmountDateStepProps {
   onNext: (amount: number, expendedAt: string) => void;
